requirements.txt
blessed==1.20.0
editor==1.6.6
inquirer==3.2.4
lxml==5.2.0
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.1
readchar==4.0.6
runs==1.2.2
six==1.16.0
tzdata==2024.1
wcwidth==0.2.13
XlsxWriter==3.2.0
xmod==1.8.1
